rbutil: fix a bug in rbsettings: FS#8512 by Sander Knopper. Also remove a warning in encodersgui.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@16185 a1c6a512-1295-4272-9138-f99709370657

+++ rbutil/rbutilqt/encodersgui.cpp
@@ -122,7 +122,7 @@ EncRbSpeexGui::EncRbSpeexGui(QDialog* parent) : QDialog(parent)
 
 }
 
-void EncRbSpeexGui::showCfg(float defQ,float defV,float defC, bool defB)
+void EncRbSpeexGui::showCfg(float defQ,float defV,int defC, bool defB)
 {
     defaultQuality =defQ;
     defaultVolume = defV;
